sprites: get it to mostly work again after large refactoring
